split translation files
