+++ src/translations/pl.ts
@@ -1,0 +1,239 @@
+export const PL = {
+  "/tickets/add": "Zgłoś potrzebę",
+  "/tickets/active": "Potrzeby",
+  "/tickets/active/mine": "Moje potrzeby",
+  "/tickets/inactive/mine": "Moje nieaktywne potrzeby",
+  "/sign-in": "Zaloguj się",
+  "/contact": "Kontakt",
+  "/faq": "FAQ",
+
+  generic: {
+    yes: "Tak",
+    no: "Nie",
+    close: "Zamknij",
+  },
+
+  pages: {
+    active: {
+      title: "Aktualne Zapotrzebowanie",
+    },
+    "sign-in": {
+      title: "Zaloguj się",
+      label: "Podaj numer telefonu",
+      placeholder: "601601601",
+      next: "Dalej",
+      "phone-verification": {
+        title: "Podaj kod weryfikacyjny",
+        label:
+          "Wysłaliśmy kod weryfikacyjny, na twój numer telefonu, w wiadomości SMS. Wpisz go poniżej aby zalogować się.",
+        placeholder: "123 456",
+        next: "Dalej",
+      },
+    },
+    main: {
+      header: "Szybkie i proste organizowanie pomocy humanitarnej",
+      subheader: "Zgłaszaj potrzeby i pomagaj!",
+      description:
+        "Powiedz czego potrzebujesz, a my przekażemy Twoją prośbę do organizacji i osób, które pomogą Ci to zorganizować.",
+      "show-all-button": "Sprawdź potrzeby",
+      "add-new-button": "Zgłoś potrzebę",
+      "steps-section": {
+        header: "Jak to działa?",
+        "step-1-title": "Zgłaszaj potrzeby",
+        "step-1-description":
+          "Wiesz, że ktoś potrzebuje transportu, noclegu lub żywności? Zgłoś tę potrzebę.",
+        "step-2-title": "My publikujemy potrzebę i szukamy rozwiązania",
+        "step-2-description":
+          "My zajmiemy się wyszukaniem organizacji lub osób, które zorganizują pomoc.",
+        "step-3-title": "Pomagaj odpowiadając na potrzeby",
+        "step-3-description":
+          "Możesz samemu pomóc? Sprawdź aktualne potrzeby i... działaj!",
+      },
+      "for-whom-section": {
+        header: "Dla kogo jest ta platforma?",
+        "list-item-1":
+          "Dla organizatorów i koordynatorów zbiórek i pomocy dla Ukrainy",
+        "list-item-2": "Dla punktów recepcyjnych na granicach",
+        "list-item-3": "Dla każdego, kto chce zorganizować pomoc!",
+      },
+    },
+    ticket: {
+      metaTitle: {
+        need: "Potrzeba",
+        cta: "Pomóż Ukrainie z potrzeby-ua.org",
+      },
+      description: {
+        "read-more": "Czytaj więcej",
+      },
+      shareButton: {
+        deliverTo: "Gdzie dostarczyć:",
+        copySuccess: "Skopiowano do schowka!",
+      },
+      verifiedOrganisation: "Zweryfikowana organizacja",
+      whoRequested: "Kto zgłosił?",
+      whoRequestedNeed: "Kto zgłosił zapotrzebowanie?",
+      whereToDeliver: "Gdzie dostarczyć?",
+      added: "Dodano",
+      share: "Udostępnij",
+      whatsNeeded: "Co potrzeba?",
+      howMuchIsNeeded: "Ile potrzeba?",
+      needActiveTill: "Zgłoszenie aktywne do",
+      views: "Odsłon",
+      needNumber: "Zgłoszenie",
+      details: "Szczegóły",
+      ticketRemovedAddNew: "Ogłoszenie usunięte. Mozesz dodać kolejne.",
+      ticketSolvedAddNew:
+        "Ogłoszenie zostało oznaczone jako rozwiąne! Mozesz dodać kolejne.",
+      ticketNotFound:
+        "Zgłoszenia nie znaleziono lub jest juz nieaktualne. Ale pewnie się coś znajdzie ",
+      here: "tutaj",
+      errorOnRemove: "Wystąpił błąd z usuwaniem zgłoszenia",
+      errorOnResolved: "Wystąpił błąd z oznaczeniem zgłoszenia jako rozwiąane",
+      warningTicketExpired: "UWAGA: Zgłoszenie nieaktualne!",
+      ticketExpiresAfterSetTime:
+        "Ogłoszenie wygasa po czasie ustalonym przez osobę zgłaszającą",
+      requesterCanExpireTicketAtAnyTime:
+        "Osoba zgłąszająca w każdym momencie może wygasić ogłoszenie",
+      lookForAnotherTicketThanksForHelp:
+        "Poszukaj innego zgłoszenia - dziękujemy za pomoc!",
+      helpType: "Rodzaj pomocy",
+      call: "Zadzwoń",
+      areYouTheAuthorOfThisTicket: "Jesteś autorem tego zgłoszenia?",
+      problemSolved: "Problem rozwiązany!",
+      remove: "Usuń",
+      needExpired: "Zgłoszenie wygasło",
+    },
+    "add-ticket": {
+      "add-need": "Dodaj potrzebę",
+      adults: "Dorosłych",
+      adultsHint: "Liczba osób",
+      adultsAge: "> 12 lat",
+      children: "Dzieci",
+      childrenHint: "Liczba dzieci",
+      childrenAge: "< 12 lat",
+      chooseLocation: "Wybierz miejsce",
+      "has-pets": "Zwierzęta",
+      "children-hint": "Liczba osób < 12 lat",
+      "what-do-you-need": "Opis potrzeby",
+      "what-do-you-need-hint":
+        "Na przykład: Materac 2-osobowy dla mamy z dzieckiem. Dlaczego i jak pilnie jest potrzebne: miałyśmy przyjąć uchodźców (mamę z dzieckiem) do naszego domu, ale ostatecznnie teraz gościmy już babcię, 3 mamy i 2 dzieci. Nie mamy gdzie ich położyć. Dodatkowe informacje: Nie mamy transportu. Prośba o przywiezienie do nas.",
+      "how-much-do-you-need": "Ile potrzebujesz?",
+      "in-pieces-if-applicable": "W sztukach, jeśli dotyczy",
+      "where-do-you-need-it-delivered": "Gdzie to potrzebujesz dostarczyć?",
+      "address-or-gps":
+        'Adres lub lokalizacja GPS (na przykład "Malczewskiego 5, Warszawa", albo "Szpital Dziecięcy w Poznaniu")',
+      "who-needs-it": "Kto to potrzebuje?",
+      "name-surname-or-org-name":
+        "Twoje imię i nazwisko lub Twoja nazwa organizacji",
+      "request-added": "Zgłoszenie przyjęte!",
+      "need-added": "Zgłoszono potrzebę!",
+      show_phone_public: "Pokaż mój numer telefonu publicznie",
+      whereFrom: "Skąd?",
+      whereTo: "Dokąd?",
+    },
+  },
+  "sign-out": "Wyloguj się",
+  "terms-of-service": {
+    title: "Regulamin",
+    "title-alternate": "Regulamin serwisu",
+    agreemenent:
+      "Zgłoszenie potrzeby jest równoznaczne z akceptacją regulaminu działania serwisu",
+  },
+  contact: {
+    "contact-us-via": "Skontaktuj się z nami przez",
+    slack: "Slack",
+    discord: "Discord",
+    github: "Github",
+    authors: "Autorzy narzędzia",
+    "in-cooperation-with": "Narzędzie zostało stworzone przy współpracy:",
+  },
+  partners: {
+    "with-us": "Współpracują z nami",
+  },
+  errors: {
+    "error-occured-while-adding": "Wystąpił błąd podczas dodawania: ",
+  },
+  auth: {
+    "you-have-been-logged-out": "Zostałeś wylogowany",
+  },
+  faq: {
+    header: "FAQ",
+    sectionOne: {
+      title: "Czym jest",
+      description:
+        "Platforma powstała w celu usprawnienia organizacji pomocy humanitarnej dla ofiar wojny w Ukrainie. W prosty sposób możesz dodać ogłoszenie dotyczące jakiegokolwiek sposobu pomocy - jeśli szukasz mieszkania dla osób z Ukrainy, potrzebujesz transportu z granicy, lub też zapewnienia jedzenia lub leków. Następnie dodajesz lokalizację - czyli np, miejsce odbioru na granicy, lub docelowe iejsce w którym szukasz zakwaaterowania - oraz numer kontaktowy do siebie. Po dodaniu informacji my zajmujemy się szukaniem organizacji lub osób, które mogą pomóc z Twoim zgłoszeniem.",
+    },
+    sectionTwo: {
+      title: "Dlaczego podaję swój numer telefonu?",
+      description:
+        "Zależy nam na sprawnym działaniu oraz szybkim rozwiązywaniu zapytań. Dlatego potrzebujemy numeru kontaktowego jedynie w celach określonych w opisie Twojej potrzeby. Używamy również numeru do potwierdzenia (poprzez wysłanie kodu SMS), że jesteś prawdziwą osobą, a nie np. botem, aby bronić się przed spamem i podobnymi aktywnościami. Twój numer nie zostaje przetrzymany ani nie będzie wykorzystany w żadnym innym celu.",
+    },
+    sectionThree: {
+      title: "Czy mogę udostępniać trwające zgłoszenia?",
+      description:
+        "Tak, możesz udostępniać wszystkie zgłoszenia na Twitterze, Telegramie oraz Facebooku, a także jako bezpośredni link.",
+    },
+    sectionFour: {
+      title:
+        "Czy mogę kontaktować się bezpośrednio z osobami szukającymi pomocy?",
+      description:
+        "Tak, podane numery służą do tego, by osoby mogące udzielić potrzebnej pomocy mogły szybko skontaktować się i zając się problemem.",
+    },
+    sectionFive: {
+      title: "Czy ogłoszenia są usuwane automatycznie?",
+      description:
+        "Tak, po 24 godzinach każde ogłoszenie zostaje automatycznie usunięte. Jeśli Twoje zgłoszenie zostało już wcześniej rozwiązane - np. udało Ci się znaleźć transport po kilku godzinach od wystawienia zgłoszenia - możesz je samodzielnie usunąć.",
+    },
+    sectionSix: {
+      title: "Gdzie znajdę regulamin serwisu?",
+      description: "Regulamin serwisu znajdziesz",
+      here: "tutaj",
+    },
+    sectionSeven: {
+      title: "Czy korzystanie z platformy jest płatne?",
+      description:
+        "Nie, platforma jest całkowicie bezpłatna. Naszym celem jest sprawne organizowanie pomocy humanitarnej.",
+    },
+    sectionEight: {
+      title: "Do czego służą kategorie?",
+      description:
+        "Kategorie pozwalają na precyzyjniejsze określenie rodzaju potrzebnej pomocy, a także sprawniejszemu wyszukiwaniu potencjalnych zgłoszeń dla osób chcących pomóc.",
+    },
+  },
+  filters: {
+    selectNeeds: "Wybierz rodzaj potrzeby",
+    all: "Wszystkie",
+    whereFrom: "Skąd?",
+    whereTo: "Dokąd?",
+  },
+  menu: {
+    open: "Open main menu",
+  },
+  metaData: {
+    title: "Pomóżmy Ukrainie!",
+    description:
+      "Szybkie i proste zarządzanie organizacją pomocy humanitarnej. Pomagasz innym? Pomożemy Ci zorganizować tę pomoc!",
+  },
+  report: {
+    title: "To zgłoszenie zawiera błędy?",
+    reportErrors: "Zgłoś nieprawidłowości",
+    actionReport: "Zgłoś",
+    actionSend: "Wyślij",
+    reason: {
+      chooseReason: "Wybierz powód",
+      options: {
+        stale: "Potrzeba nieaktualna",
+        lacksContact: "Brakuje danych kontaktowych",
+        lacksDescription: "Brakuje informacji",
+        itsAnOffer: "To oferta, a nie potrzeba",
+        itsAnAdvertisementOrSpam: "To reklama albo spam",
+        wrongCategory: "Nieprawidłowa kategoria",
+        other: "Inny powód",
+      },
+    },
+    reasonDetails: "Podaj więcej informacji",
+    thankYou: "Dziękujemy!",
+    thankYouForReport:
+      "Dziękujemy! Twoje zgłoszenie pozwoli nam usprawnić udzielanie pomocy. Nasi moderatorzy za chwilę się temu przyjrzą.",
+  },
+}
